Update enum debuginfo tests
Bug #52452 notes some debuginfo test regressions when moving to gdb
8.1.  This series will also cause versions of gdb before 8.2 to fail
when a recent LLVM is used -- DW_TAG_variant_part support was not
added until 8.2.

This patch updates one of the builders to a later version of Ubuntu,
and then arranges to install gdb 8.2 from the toolchain PPA.  It
updates the relevant tests to require both a new-enough LLVM and a
new-enough gdb; the subsequent patch arranges to continue testing the
fallback mode.

The "gdbg" results are removed from these tests because the tests now
require a rust-enabled gdb.